+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The following is a curated list of changes in the Enact project, newest changes 
 - `core/util` function `checkPropTypes` to check the prop types of a component
 - `spotlight/SpotlightRootDecorator` config `focusEffectClass` and function `setFocusEffectClass` to apply a css class to every spottable component when it receives spotlight focus.
 - `ui/resolution` Added support for linear scaling type
+- `ui/Slider` support for `colorPicker`
 
 ### Changed